First attempt using chat gpt
basic upload

I'd like to start a tool that creates network graphs in streamlit.

Step one - let's provide an option for a user to upload a csv. There should be a dropdown to select a column in the csv that is the Source and an option to select a target.

In terms of the upload I'd like for the user, after uploading the csv, to see a preview of the first 50 rows of the csv. They should be prompted to provide how many rows they'd like to skip - because some csvs we'll need to skip a few rows at the start.

The file uploaded should be saved as 'df', read into pandas.

I'd like to include print notifications at big step in the process to help debugging.