fixed reading header
See these two screenshots. In the first, after uploading a file that includes several poorly formatted rows, we get an error - "Error reading file: Error tokenizing data. C error: Expected 2 fields in line 7, saw 145"

After we select the right number of rows to skip, in this case 5, we get a good preview. Is there a way to read such a csv for the preview so it can handle cases where there are poorly formatted header rows or not?

If not, can you suggest a good design to handle this?